feat(ui-hooks): rename filename
