+++ webapp/src/hooks/useBlocker.ts
@@ -15,7 +15,7 @@
 import type { History, Transition } from "history";
 import { useContext, useEffect } from "react";
 import { UNSAFE_NavigationContext as NavigationContext } from "react-router-dom";
-import useAutoUpdateRef from "./useAutoUpdateRef";
+import useAutoUpdateRef from "./useUpdatedRef";
 
 // * Workaround until it will be supported by react-router v6.
 // * Based on https://ui.dev/react-router-preventing-transitions
