+++ webapp/src/hooks/useSafeMemo.ts
